requirements.txt
#pip install -r requirements.txt to get all packages

# Django
Django>=5.1.0

# ML/AI Libraries
#unccoment line below for torch with cuda support
#--extra-index-url https://download.pytorch.org/whl/cu126
torch>=2.5.0

# Other ML libraries
scikit-learn>=1.6.0
pandas>=2.2.0

# Document Processing
python-docx>=1.1.0
lxml>=5.0.0

# Training & Evaluation
tensorboard>=2.18.0
tqdm>=4.67.0

# API & Serialization
djangorestframework>=3.15.0

# Utilities
joblib>=1.4.0

# Additional useful packages
huggingface-hub>=0.26.0
safetensors>=0.4.0
